+++ plugin.php
@@ -10,7 +10,7 @@
  * Plugin Name: Rollback Update Failure
  * Author: WP Core Contributors
  * Description: Feature plugin to test plugin/theme update failures and rollback to previous installed packages.
- * Version: 4.0.0
+ * Version: 4.1.0
  * Network: true
  * License: MIT
  * Text Domain: rollback-update-failure
@@ -33,38 +33,19 @@ if ( ! defined( 'WPINC' ) ) {
 // Load the Composer autoloader.
 require __DIR__ . '/vendor/autoload.php';
 
-add_action(
-	'plugins_loaded',
-	function() {
-		\WP_Dependency_Installer::instance( __DIR__ )->run();
-	}
-);
-
-// TODO: Deactivate plugin when/if committed to core.
-// if ( version_compare( get_bloginfo( 'version' ), '6.3-beta1', '>=' ) ) {
-// require_once ABSPATH . 'wp-admin/includes/plugin.php';
-// deactivate_plugins( __FILE__ );
-// }
-
 // Add to wp-admin/includes/admin-filters.php.
 add_action( 'init', array( 'WP_Rollback_Auto_Update', 'init' ) );
 
-// Hopefully add some VirtualBox compatibility.
 add_action(
-	'post_move_dir',
+	'plugins_loaded',
 	function() {
-		/*
-		 * VirtualBox has a bug when PHP's rename() is followed by an unlink().
-		 *
-		 * The bug is caused by delayed clearing of the filesystem cache, and
-		 * the solution is to clear dentries and inodes at the system level.
-		 *
-		 * Most hosts add shell_exec() to the disable_function directive.
-		 * function_exists() is usually sufficient to detect this.
-		 */
-		if ( function_exists( 'shell_exec' ) ) {
-			// phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.system_calls_shell_exec
-			shell_exec( 'sync; echo 2 > /proc/sys/vm/drop_caches' );
+		if ( ! \is_plugin_active( 'faster-updates/faster-updates.php' ) ) {
+			echo '<div class="error notice is-dismissible"><p>';
+			print(
+				wp_kses_post( __( '<strong>Rollback Update Failure</strong> cannot run unless the <strong>Faster Updates</strong> plugin is active. Please refer to the readme.', 'rollback-update-failure' ) )
+			);
+			echo '</p></div>';
+			\deactivate_plugins( __FILE__ );
 		}
 	}
 );